ExactlyEnv to be deprecated
* `ExactlyEnv` was still creating the `DefaultEnv`
* We moved the static `create()` method `DefaultEnv` and removed
  `ExactlyEnv` from exactlyUtils
